requirements.txt
python-telegram-bot==21.3
ccxt==4.4.20
pandas==2.2.3
numpy==2.1.3
aiohttp==3.11.7
aiosqlite==0.20.0
